[#3] Fix tests, requirements, flake8 and coverage
