+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include requirements.txt
+include requirements-test.txt
