+++ photonpump/exceptions.py
@@ -1,9 +1,9 @@
 class PhotonPumpException(Exception):
     pass
 
+
 class StreamNotFoundException(PhotonPumpException):
 
     def __init__(self, message, stream):
         super().__init__(message)
         self.stream = stream
-
