+++ requirements-test.txt
@@ -1,0 +1,5 @@
+pytest
+pytest-asyncio
+pytest-cov
+colorama
+flake8
