+++ setup.cfg
@@ -1,0 +1,14 @@
+[aliases]
+test = pytest
+
+[tool:pytest]
+addopts = --verbose --color=yes --cov=photonpump
+
+[flake8]
+exclude = __pycache__,photonpump/messages_pb2.py,.eggs
+
+[coverage:run]
+branch = True
+
+[coverage:report]
+show_missing = on
