+++ test/write_test.py
@@ -1,8 +1,10 @@
+import json
+import uuid
+
+import pytest
+
 from photonpump import connect, messages, messages_pb2
 
-import json
-import pytest
-import uuid
 
 @pytest.mark.asyncio
 async def test_single_event_publish(event_loop):
@@ -22,6 +24,7 @@ async def test_single_event_publish(event_loop):
 
         assert isinstance(result, messages_pb2.WriteEventsCompleted)
         assert result.first_event_number == 0
+
 
 @pytest.mark.asyncio
 async def test_three_events_publish(event_loop):
@@ -51,6 +54,7 @@ async def test_three_events_publish(event_loop):
         assert result.first_event_number == 0
         assert result.last_event_number == 2
 
+
 @pytest.mark.asyncio
 async def test_a_large_event(event_loop):
 
@@ -60,10 +64,13 @@ async def test_a_large_event(event_loop):
         data = json.load(f)
 
     async with connect(loop=event_loop) as c:
-        write_result = await c.publish(stream_name, [messages.NewEvent('big_json', data=data),
-                                                     messages.NewEvent('big_json', data=data),
-                                                     messages.NewEvent('big_json', data=data)
-            ])
+        write_result = await c.publish(
+            stream_name, [
+                messages.NewEvent('big_json', data=data),
+                messages.NewEvent('big_json', data=data),
+                messages.NewEvent('big_json', data=data)
+            ]
+        )
         assert write_result.first_event_number == 0
         read_result = await c.get(stream_name, 0)
         assert read_result[0].type == 'big_json'
